backend/Python/features/ai_config.py
import google.generativeai as genai
from dotenv import load_dotenv
import os

# Load env vars
load_dotenv()
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")

genai.configure(api_key=GOOGLE_API_KEY)

model = genai.GenerativeModel("gemini-2.5-flash")
