backend/Python/requirements.txt
fastapi[standard]
uvicorn[standard]
google-generativeai
python-dotenv
googletrans==4.0.0-rc1
google-cloud-translate